refactor(tests): split share code out into util_test, rename main tests
